.OpenServer class

The OpenServer class contains functions that communicate with Petroleum Experts applications such as Prosper, GAP and MBAL. The function send commands, retrieve data and write data to OpenServer variables exposed by the Windows COM interface.

Create a new OpenServer instance.

Arguments

server
passing an OpenServer object if available

variable
OpenServer variable

set_value
value to set to OpenServer variable

command
an OpenServer command

Gv
GAP variable

Value

An OpenServer object.

OpenServer class

.OpenServer

OpenServerR6

Constructors

Constructor

newOpenServer()

setOpenServer()

openserver()

Public Methods

Parameters Methods	Action

DoCmd()	Send command

DoGet()	Retrieve value from variable

DoSet()	Set value to variable

DoSlowCmd()	Send slow command

DoGAPFunc()	Run GAP function

R6 Usage

$DoCmd(command)

$DoGet(variable)

$DoSet(variable, set_value)

$DoSlowCmd(command)

$DoGAPFunc(Gv)

S3 Usage

DoCmd(server, command)

DoGet(server, variable)

DoSet(server, variable, set_value)

DoSlowCmd(server, command)

DoGAPFunc(server, Gv)

See also